requirements.txt
langchain
langchain-community
langchain-groq
langgraph
faiss-cpu
sentence-transformers
streamlit
python-dotenv
sqlalchemy
pandas
groq